fix: 사이즈 기준 변경 및 쿼리, 변수 일부 수정

+++ potfill/src/main/java/com/potfill/admin/district/model/Complaint.java
@@ -15,7 +15,7 @@ import lombok.ToString;
 @AllArgsConstructor
 @Builder
 public class Complaint {
-	private int complaintId;
+	private long complaintId;
 	private String reporterName;
 	private String reporterNumber;
 	private String incidentAddress;
